// src/channels/model/OutputFileColumnField.ts
enum OutputFileColumnField {
  PRODUCT_ID = 'productId',
  INVENTORY_ITEM_VALUE = 'inventoryItemValue',
  SKU = 'sku',
  UPC = 'upc',
  PRODUCT_TITLE = 'productTitle',
  PRODUCT_DESCRIPTION = 'productDescription',
  PRODUCT_PRIMARY_CATEGORY = 'productPrimaryCategory',
  PRODUCT_SECONDARY_CATEGORY = 'productSecondaryCategory',
  PRODUCT_VENDER = 'productVendor',
  OPTION_VALUE = 'optionValue',
  PRODUCT_IMAGE = 'productImage',
  PRODUCT_TAGS = 'productTags',
  TITLE = 'title',
  DESCRIPTION = 'description',
  PRIMARY_CATEGORY = 'primaryCategory',
  SECONDARY_CATEGORY = 'secondaryCategory',
  VENDOR = 'vendor',
  QUANTITY = 'quantity',
  WEIGHT = 'weight',
  ALPHABETIC_CURRENCY_CODE = 'alphabeticCurrencyCode',
  RETAIL_PRICE = 'retailPrice',
  WHOLESALE_PRICE = 'wholesalePrice',
  CONDITION = 'condition',
  IMAGE = 'image',
  TAGS = 'tags',
  ADDITIONAL_FIELD_VALUE = 'additionalFieldValue'
}
export default OutputFileColumnField;